Created TextFileParser and EssayManager
TextFileParser takes a text file (that has the entire essay in it) and parses it for a key character that indicates BREAKS and when research is needed (default character is ">"). It breaks the essay up into sections according to this symbol and has a public method that can be used to get the section at a given index value.

The EssayManager can truncate an essay section into subsections according to a given # of characters per keystroke. Then, if the player can type and is on the essay screen, they can press any input except for keys that are otherwise used and it updates a TMPro asset subsection by subsection until it reaches the end of the subsections, at which point research is needed.

*Currently* no signal is sent elsewhere to indicate that the end of a section has been reached. It also has the check for correct game state commented out because the game state requires a game manager with proper connections to many other managers that I don't have set up in my test scene and I don't want to mess with Sam's!